alacritty導入

+++ setup.sh
@@ -4,7 +4,12 @@ if [ ! -d ~/.config ]; then
   mkdir ~/.config
 fi
 
-files=("vimrc" "tmux.conf" "config/git" "config/nvim" "config/efm-langserver" "vim")
+if [ ! -d ~/.config/alacritty/themes ]; then
+  mkdir -p ~/.config/alacritty/themes
+  git clone https://github.com/alacritty/alacritty-theme ~/.config/alacritty/themes
+fi
+
+files=("vimrc" "tmux.conf" "config/git" "config/nvim" "config/alacritty/alacritty.toml" "config/efm-langserver" "vim")
 for file in "${files[@]}"
 do
    if [ ! -e ~/.$file ];then
@@ -12,5 +17,7 @@ do
    fi
 done
 
-ln -s ~/.vim/autoload ~/.config/nvim/autoload
-ln -s ~/.vim/denops ~/.config/nvim/denops
+if [ ! -e ~/.config ]; then
+  ln -s ~/.vim/autoload ~/.config/nvim/autoload
+  ln -s ~/.vim/denops ~/.config/nvim/denops
+fi
